Use products related to the last SIX messages
This allows for better multi-turn conversations.
For instance, the following worked:
User: Do you sell purple sneakers?
Assistant: Yes, we do! We have a pair of lavender canvas sneakers that are a fresh take on classic casual footwear. They are a low-top silhouette with a traditional lace-up closure and a cushioned footbed for comfort. Would you like to know more about them?
User: How much are they?
Assistant: The lavender canvas sneakers are $60.